Removing old sync files
