--- app/src/main/java/org/qosp/notes/data/sync/core/ISyncBackend.kt
@@ -1,21 +1,0 @@
-package org.qosp.notes.data.sync.core
-
-import org.qosp.notes.data.model.IdMapping
-import org.qosp.notes.data.model.Note
-import org.qosp.notes.preferences.CloudService
-
-interface ISyncBackend<T : ProviderConfig, N: SyncNote> {
-    val service: CloudService
-
-    suspend fun getSyncNoteFrom(note: Note, idMapping: IdMapping): N
-
-    //    suspend fun sync(config: T): BaseResult
-    suspend fun list(config: T): SyncResult<List<N>>
-    suspend fun createNote(note: Note, config: T): SyncResult<N>
-    suspend fun getNoteContent(note: N, config: T): SyncResult<N>
-    suspend fun deleteNote(note: N, config: T): SyncResult<Unit>
-    suspend fun updateNote(note: N, config: T): SyncResult<Unit>
-
-    suspend fun authenticate(config: T): BaseResult
-    suspend fun isServerCompatible(config: T): BaseResult
-}
