--- app/src/main/java/org/qosp/notes/data/sync/core/ProviderConfig.kt
@@ -1,7 +1,0 @@
-package org.qosp.notes.data.sync.core
-
-import org.qosp.notes.preferences.CloudService
-
-interface ProviderConfig {
-    val provider: CloudService
-}
