--- app/src/main/java/org/qosp/notes/data/sync/core/SyncResult.kt
@@ -1,6 +1,0 @@
-package org.qosp.notes.data.sync.core
-
-sealed class SyncResult<T> {
-    data class Success<T>(val data: T) : SyncResult<T>()
-    data class Error<T>(val error: Throwable) : SyncResult<T>()
-}
