+++ app/src/main/java/org/qosp/notes/data/sync/fs/StorageConfig.kt
@@ -4,15 +4,9 @@ import android.net.Uri
 import androidx.core.net.toUri
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
-import org.qosp.notes.data.sync.core.ProviderConfig
-import org.qosp.notes.preferences.CloudService
 import org.qosp.notes.preferences.PreferenceRepository
 
-data class StorageConfig(
-    val location: Uri,
-    override val provider: CloudService = CloudService.FILE_STORAGE
-) :
-    ProviderConfig {
+data class StorageConfig(val location: Uri) {
 
     companion object {
         fun storageLocation(prefRepo: PreferenceRepository): Flow<StorageConfig?> =
