+++ app/src/main/java/org/qosp/notes/data/sync/neu/NewSyncNote.kt
@@ -1,7 +1,5 @@
 package org.qosp.notes.data.sync.neu
 
-import org.qosp.notes.data.sync.core.NextcloudNote
-
 data class NewSyncNote(
     val id: Long,
     val idStr: String,
@@ -12,34 +10,4 @@ data class NewSyncNote(
     val category: String = "",
     val favorite: Boolean = false,
     val readOnly: Boolean = false,
-) {
-    fun asNextCloudNote(): NextcloudNote = NextcloudNote(
-        id = id,
-        content = content,
-        title = title,
-        modified = lastModified,
-        etag = extra,
-        category = category,
-        favorite = favorite,
-        readOnly = readOnly,
-        remoteId = "",
-    )
-}
-
-fun NextcloudNote.asNewSyncNote(): NewSyncNote = NewSyncNote(
-    id = id,
-    idStr = "",
-    content = content,
-    title = title,
-    lastModified = modified,
-    extra = etag,
-    category = category,
-    favorite = favorite,
-    readOnly = readOnly == true,
-)
-
-data class RemoteNoteMetaData(
-    val id: String,
-    val title: String,
-    val lastModified: Long,
 )
