+++ app/src/main/java/org/qosp/notes/data/sync/neu/RemoteNoteMetaData.kt
@@ -1,0 +1,7 @@
+package org.qosp.notes.data.sync.neu
+
+data class RemoteNoteMetaData(
+    val id: String,
+    val title: String,
+    val lastModified: Long,
+)
