+++ app/src/main/java/org/qosp/notes/data/sync/nextcloud/NextcloudNote.kt
@@ -1,0 +1,16 @@
+package org.qosp.notes.data.sync.nextcloud
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class NextcloudNote(
+    val id: Long,
+    val etag: String? = null,
+    val content: String?,
+    val title: String,
+    val category: String,
+    val favorite: Boolean,
+    val modified: Long,
+    val readOnly: Boolean? = null,
+    val remoteId: String = id.toString(),
+)
