+++ app/src/main/java/org/qosp/notes/data/sync/nextcloud/model/NextcloudConverters.kt
@@ -1,7 +1,7 @@
 package org.qosp.notes.data.sync.nextcloud.model
 
 import org.qosp.notes.data.model.Note
-import org.qosp.notes.data.sync.core.NextcloudNote
+import org.qosp.notes.data.sync.nextcloud.NextcloudNote
 
 fun Note.asNextcloudNote(id: Long, category: String): NextcloudNote = NextcloudNote(
     id = id,
@@ -10,13 +10,4 @@ fun Note.asNextcloudNote(id: Long, category: String): NextcloudNote = NextcloudN
     category = category,
     favorite = isPinned,
     modified = modifiedDate
-)
-
-fun NextcloudNote.asNewLocalNote(id: Long, notebookId: Long?) = Note(
-    id = id,
-    title = title,
-    content = content ?: "",
-    isPinned = favorite,
-    modifiedDate = modified,
-    notebookId = notebookId
 )
